Move utils file to common
