+++ model_compression_toolkit/trainable_infrastructure/common/util.py
